Strip NDWarp from its core processing. Add simple proof of concept.
NDWarp now doesnt warps anything.
Its processCallbacks has been stripped of warping capabilities. Now, the only thing it
does is setting the first 4718592 bytes (half of an 8-bytes-per-pixel image of 3072x3072 pixels)
to 1.

So we have a proof of concept of something that receives an array, processes it and returns it back
to plugins.

Its kinda ready to receive SSCPiMega API, but we still need to clean LOTS of stuff:

    - old .db files are probably mainly useless now.
    - ndutil.h is probably useless now.
    - asyn writeInt and WriteFloat tc. functions probably need to be cleaned too.
    - Everything thats called NDWarp should be called NDSSCPiMega or something similar.